feat: custom error생성 및 review에 적용

+++ cake-quake-back/src/main/java/com/cakequake/cakequakeback/review/service/buyer/BuyerReviewServiceImpl.java
@@ -2,11 +2,12 @@ package com.cakequake.cakequakeback.review.service.buyer;
 
 import com.cakequake.cakequakeback.common.dto.InfiniteScrollResponseDTO;
 import com.cakequake.cakequakeback.common.dto.PageRequestDTO;
+import com.cakequake.cakequakeback.common.exception.BusinessException;
+import com.cakequake.cakequakeback.common.exception.ErrorCode;
 import com.cakequake.cakequakeback.review.dto.ReviewRequestDTO;
 import com.cakequake.cakequakeback.review.dto.ReviewResponseDTO;
 import com.cakequake.cakequakeback.review.entities.Review;
 import com.cakequake.cakequakeback.review.repo.buyer.BuyerReviewRepo;
-import com.cakequake.cakequakeback.review.repo.common.CommonReviewRepo;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.data.domain.Page;
@@ -36,11 +37,11 @@ public class BuyerReviewServiceImpl implements BuyerReviewService {
 
         //해당 주문서가 존재하지 않거나 권한이 없는 경우
 //        CakeOrder order = orderRepo.findByIdAndUserId(orderId,userId)
-//        .orElseThrow(()-> new BusinessException(1008, "해당 주문 정보를 찾을 수 없거나 권한이 없습니다."));
+//        .orElseThrow(() -> new BusinessException(ErrorCode.NOT_FOUND_ORDER_ID));
 
         //이미 리뷰 작성 여부 체크
 //        if(reviewRepo.findByOrderId(orderId).isPresent()){
-//            throw new BussinessException(708, "해당 주문에는 이미 리뷰가 작성 되었습니다.");
+//            throw new BusinessException(ErrorCode.ALREADY_REVIEWED_ORDER);
 //        }
 
         //DTO -> Entity 매핑
@@ -82,30 +83,30 @@ public class BuyerReviewServiceImpl implements BuyerReviewService {
 
     //구매자 리뷰 단건 조회
     @Override
-    public ReviewResponseDTO getReview(Long reviewId, Long userId) {
+    public ReviewResponseDTO getReview(Long reviewId, Long uid) {
         //리뷰 존재 여부 체크
         Review review = buyerReviewRepo.findById(reviewId)
-                .orElse(() -> new BussinessExcepion(1004, "해당 리뷰를 찾을 수 없습니다"));
-//
-//        //본인 리뷰인지 검증
-//        if (!review.getUser().getUserId().equals(userId)) {
-//            throw new BussinessException(906, "본인 리뷰만 조회할 수 있습니다");
-//        }
+                .orElseThrow(() -> new BusinessException(ErrorCode.NOT_FOUND_REVIEW_ID));
+        //본인 리뷰인지 검증
+        if (!review.getUser().getUid().equals(uid)) {
+            throw new BusinessException(ErrorCode.NOT_AUTHORIZED_OTHER);
+        }
 
         return buyerReviewRepo.selectDTO(reviewId);
 
     }
     //구매자가 자신 리뷰 수정
     @Override
-    public ReviewResponseDTO updateReview(Long reviewId, ReviewRequestDTO dto, Long userId) {
-        //        //리뷰 존재 여부 체크
-       Review review = buyerReviewRepo.findById(reviewId)
-                .orElse(() -> new BussinessExcepion(1004, "해당 리뷰를 찾을 수 없습니다"));
-//
-//        //본인 리뷰인지 검증
-//        if (!review.getUser().getUserId().equals(userId)) {
-//            throw new BussinessException(906, "본인 리뷰만 조회할 수 있습니다");
-//        }
+    public ReviewResponseDTO updateReview(Long reviewId, ReviewRequestDTO dto, Long uid) {
+        //리뷰 존재 여부 체크
+        Review review = buyerReviewRepo.findById(reviewId)
+                .orElseThrow(() ->
+                        new BusinessException(ErrorCode.NOT_FOUND_REVIEW_ID)
+                );
+        //본인 리뷰인지 검증
+        if (!review.getUser().getUid().equals(uid)) {
+            throw new BusinessException(ErrorCode.NOT_AUTHORIZED_OTHER);
+        }
 
         //수정 가능한 필드만 수정하기
         review.setRating(dto.getRating());
@@ -119,15 +120,16 @@ public class BuyerReviewServiceImpl implements BuyerReviewService {
 
     //리뷰 삭제
     @Override
-    public void deleteReview(Long reviewId, Long userId) {
-        //        //리뷰 존재 여부 체크
-         Review review = buyerReviewRepo.findById(reviewId)
-              .orElse(() -> new BussinessExcepion(1004, "해당 리뷰를 찾을 수 없습니다"));
-//
-//        //본인 리뷰인지 검증
-//        if (!review.getUser().getUserId().equals(userId)) {
-//            throw new BussinessException(906, "본인 리뷰만 조회할 수 있습니다");
-//        }
+    public void deleteReview(Long reviewId, Long uid) {
+        //리뷰 존재 여부 체크
+        Review review = buyerReviewRepo.findById(reviewId)
+                .orElseThrow(() ->
+                        new BusinessException(ErrorCode.NOT_FOUND_REVIEW_ID)
+                );
+        //본인 리뷰인지 검증
+        if (!review.getUser().getUid().equals(uid)) {
+            throw new BusinessException(ErrorCode.NOT_AUTHORIZED_OTHER);
+        }
 
         //삭제 -> 상태만 DELETE로 변경
         review.deleteByBuyer();
